requirements.txt
flask
gunicorn
openai
python-dotenv
werkzeug
google-generativeai
numpy

